feat : added lint script in forge

+++ packages/foundry/src/YourContract.sol
@@ -18,15 +18,10 @@ contract YourContract {
     string public greeting = "Building Unstoppable Apps!!!";
     bool public premium = false;
     uint256 public totalCounter = 0;
-    mapping(address => uint) public userGreetingCounter;
+    mapping(address => uint256) public userGreetingCounter;
 
     // Events: a way to emit log statements from smart contract that can be listened to by external parties
-    event GreetingChange(
-        address indexed greetingSetter,
-        string newGreeting,
-        bool premium,
-        uint256 value
-    );
+    event GreetingChange(address indexed greetingSetter, string newGreeting, bool premium, uint256 value);
 
     // Constructor: Called once on contract deployment
     // Check packages/foundry/deploy/Deploy.s.sol
@@ -74,7 +69,7 @@ contract YourContract {
      * The function can only be called by the owner of the contract as defined by the isOwner modifier
      */
     function withdraw() public isOwner {
-        (bool success, ) = owner.call{value: address(this).balance}("");
+        (bool success,) = owner.call{value: address(this).balance}("");
         require(success, "Failed to send Ether");
     }
